Moving components to appropriate folder
