+++ react/components/GalleryLayoutItem.tsx
